generalized logic into scraper class that uses config dicts to get jobs
Previously the scraping logic was done more-or-less the same in four
different functions; now there's one object that looks for all the same
information.  Where it looks (from the URL to the CSS selectors) is
defined in a key word dictionary that gets passed to the scraper object.

Doing new searches is now as simple as creating new entries in the
configs object, defining everything that's passed to the scraper.