Converted config dicts to proper class, implemented @dataclass
Instead of having a key word dictionary for every search to do, now
there's a class for the configs that represents each of the job sites.
To do a search, you add the term to a list at the top of the file and it
will iterate through the terms for each of the websites.

I also implemented the @dataclass decorator across all the classes to
test out the size and speediness of the implementation.  There wasn't
much size difference between the kw dicts and the classes until
__slots__ were added; that almost halved the size of each object.  It
also seemed to run faster, but that's just anecdotal.  I added timing
code that /seemed/ to improve, but it's hard to tell without lots of
repeated running and stats because of the multithreading and network IO.