Reworked the whole architecture to better utilized tor and threading
Tor wasn't being utilized well, and it's inefficient to try to create a
new circuit for each request.  So to avoid being flagged as a bot, my
solution is to create one thread for each search / website combination,
and to add a sleep between requesting each of the detail pages to make
it look more like a real person browsing.

I also added more ports for tor to listen on, so that multiple searches
can be run in parallel to the same website from different IPs.

So in theory, it should create 16 threads to run in parallel (four
searches across four websites) with each search term getting its own IP
address to utilize for the scraper.